Revamp Spire integration so that we invoke it on whole programs.
There is still some leftover code in places, but the basic idea here is that we now invoke Spire at the point where we create a `ProgramVersion` and pass it all of the individual shaders at once.
The `Shader` creation APIs now expect to work with raw native-API shader code, and don't even provide D3D reflection information back to the user.

The reflection interface also needed to be updated to assume it gets a single `spire::ShaderReflection*` instead of one per stage. This mostly just simplifies the logic, but eventually I'll need to make sure that we update Spire to be able to produce per-stage masks in cases where that can help performance.

ModelViewer runs after these changes, and all the Spire regression tests pass, except for `error0.hlsl`, which is currently an expected failure.